requirements.txt
moviepy==2.1.2
google-cloud-texttospeech==2.15.1
matplotlib==3.9.4
numpy==2.0.2